requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2025.6
Django==5.2.3
django-cleanup==9.0.0
django-crispy-forms==2.4
django-jazzmin==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
